refactor: single used components moved to SendToWalletPage and deleted the components
