--- packages/widget/src/components/CustomWalletBookmarkSwitch/index.ts
@@ -1,1 +1,0 @@
-export * from './CustomWalletBookmarkSwitch';
