--- packages/widget/src/components/SendToWallet/index.ts
@@ -1,1 +1,0 @@
-export * from './SendToWallet';
