--- packages/widget/src/components/SendToWalletConfirmButton/index.ts
@@ -1,1 +1,0 @@
-export * from './SendToWalletConfirmButton';
