+++ packages/widget/src/pages/SendToWalletPage/SendToWalletPage.styled.tsx
@@ -1,10 +1,27 @@
+import {
+  Alert,
+  Container,
+  InputBase,
+  FormControl as MuiFormControl,
+} from '@mui/material';
+import { inputBaseClasses } from '@mui/material/InputBase';
 import { styled } from '@mui/material/styles';
-import { Alert, Container } from '@mui/material';
 
-export const PageContainer = styled(Container)(({ theme }) => ({
-  height: 508,
-  display: 'flex',
-  flexDirection: 'column',
+export const FormControl = styled(MuiFormControl)(({ theme }) => ({
+  padding: theme.spacing(1.5, 2, 1.5, 0),
+}));
+
+export const Input = styled(InputBase)(({ theme }) => ({
+  [`.${inputBaseClasses.input}`]: {
+    height: 32,
+    padding: theme.spacing(0, 0, 0, 2),
+  },
+  [`&.${inputBaseClasses.disabled}`]: {
+    color: 'inherit',
+  },
+  [`.${inputBaseClasses.input}.${inputBaseClasses.disabled}`]: {
+    WebkitTextFillColor: 'unset',
+  },
 }));
 
 export const AlertSection = styled(Alert)(({ theme }) => ({
@@ -15,4 +32,10 @@ export const AlertSection = styled(Alert)(({ theme }) => ({
     color: theme.palette.grey[700],
   },
   marginBottom: theme.spacing(2),
+}));
+
+export const PageContainer = styled(Container)(({ theme }) => ({
+  height: 508,
+  display: 'flex',
+  flexDirection: 'column',
 }));
